Remove obsolete metrics code
